move login and signup check to welcome template
